requirements.txt
amqp==5.1.0
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
black==22.1.0
blinker==1.4
Brotli==1.0.9
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.5.0
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
Flask-WTF==1.0.0
flower==1.0.0
fonttools==4.31.2
greenlet==1.1.2
html5lib==1.1
humanize==4.0.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.12.0
pytz==2021.3
redis==4.2.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
typing-extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.2
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
zopfli==0.2.1
